Adjust singleton load order, simulation scripts, have InventoryManager generate products itself
