+++ JungleGo/Assets/_Scripts/Systems/InventoryManager.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
@@ -8,6 +9,8 @@ public class InventoryManager : MonoBehaviour
     public static InventoryManager Instance { get; private set; }
 
     public Dictionary<Product, int> Stock { get; private set; }
+    private List<int> uniqueProductIds { get;set; }
+    private System.Random random { get; set; }
 
     private void Awake()
     {
@@ -16,6 +19,7 @@ public class InventoryManager : MonoBehaviour
             Instance = this;
             // Persist this object across scenes
             DontDestroyOnLoad(gameObject);
+            random = new System.Random();
         }
         else
         {
@@ -23,14 +27,14 @@ public class InventoryManager : MonoBehaviour
         }
     }
 
-    public void GenerateStock(List<Product> products, int maxProductCount = Constants.MaxUniqueProducts)
+    public void GenerateStock(int maxProductCount = Constants.MaxUniqueProducts)
     {
         var newStock = new Dictionary<Product, int>();
-        var db = ProductDatabase.Instance;
+        uniqueProductIds = new List<int>();
 
-        foreach (var product in products)
-        {
-            var count = (int)System.Math.Ceiling(Random.value * maxProductCount); // ensure there is at least 1 product
+        for (int i = 0; i < Constants.MaxUniqueProducts; i++) {
+            var product = PickProduct();
+            var count = (int)System.Math.Ceiling(UnityEngine.Random.value * maxProductCount); // ensure there is at least 1 product
             newStock.Add(product, count);
             Debug.Log($"Inventory: Added {count} {product.ProductName}s to inventory");
         }
@@ -55,9 +59,28 @@ public class InventoryManager : MonoBehaviour
         return false;
     }
 
-    public Product GetRandomProduct()
+    /// <summary>
+    /// Select a random unique product from our database if possible, otherwise it'll refresh the list of products and reuse them
+    /// </summary>
+    private Product PickProduct()
     {
-        var index = Random.Range(0, Stock.Count);
+        // Refresh products list if we run out
+        if (uniqueProductIds.Count <= 0) {
+            for (int i = 0; i < ProductDatabase.Instance.ProductCount; i++) {
+                uniqueProductIds.Add(i + 1);
+            }
+        }
+
+        int index = random.Next(0, uniqueProductIds.Count);
+        var id = uniqueProductIds[index];
+        uniqueProductIds.RemoveAt(index);
+
+        return ProductDatabase.Instance.GetProductById(id);
+    }
+
+    public Product GetRandomProductFromInventory()
+    {
+        var index = UnityEngine.Random.Range(0, Stock.Count);
         return Stock.Keys.ElementAt<Product>(index);
     }
 }
